refactor : 입력 로직에서 도메인 로직 분리
