+++ README.md
@@ -1,35 +1,16 @@
 # 기능목록
 
-- 입력 : `경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).`
-  - 쉼표로 구분하여 배열에 저장
-  - 유효성 검사
-    - 5자 이하
-    - 쉼표를 기준으로 구분
-  - (예외) 5자 초과
-  - (예외) 0자
-  - 실패했을경우 돌아감
-- 입력 : `시도할 횟수는 몇 회인가요?`
-
-  - 유효성 검사
-    - 양의 정수
-  - (예외) 양의 정수가 아닌 값
-  - 실패했을경우 돌아감
-
-- 출력 : `실행 결과`
-- (반복)현재 게임 상황 출력
-  - 각차마다 0부터 9사이의 무작위 값을 구함.
-  - 그 값이 4이상이라면 전진한다. 아니라면 정지한다.
-  ```
-  pobi : -
-  woni :
-  jun : -
-  ```
-- 최종우승자 출력
-  - 우승자를 쉼표로 나누어 출력(여러명일 경우)
-  ```
-  최종 우승자: pobi, jun
-  ```
-
+- 사용자에게서 자동차 이름을 입력받는다. (ui)
+- 사용자에게서 경주 횟수를 입력받는다 (ui)
+- 입력받은 이름으로 자동차를 만든다 (domain) - Car
+  - 이름은 5자 이하만 가능하다 (domain)
+- 입력받은 횟수만큼 자동차 경주게임을 한다 - Race
+  - 자동차 경주게임 한 라운드를 진행한다. (domain)
+  - 게임 한 라운드 진행상황을 출력해준다 (ui)
+- 자동차 경주게임이 끝나면 우승자를 알려준다 - Winner
+  - 우승자를 판단할수있다. (domain)
+  - 우승자는 한명이상일수도있다 (domain) 
+  - 우승자가 여러명이면 쉼표로 구분해서 출력해준다 (ui)
 
 
 # 피드백으로부터 수정한 부분
@@ -42,6 +23,13 @@
 
 # 피드백으로부터 공부할 부분
 - .reduce, .filter .map등 js에 있는 유용한 문법에 대해서 모르는 경향이 있다. 따라서 해당 부분에 대해서 공부할 필요가 있음.
+
+
+# 2단계 미션 요구사항
+- [ ] 도메인 로직은 domain/ 하위로, UI 관련 로직은 view/ 하위에서 관리한다.
+- [ ] domain/ 하위의 모듈은 view/ 하위의 모듈을 의존하지 않아야 한다.
+- [ ] 도메인 로직 내에서 테스트하기 어려운 부분을 분리하고, 테스트 가능한 부분에 대해서만 단위 테스트를 작성한다.
+- [ ] 테스트 코드에서 jest.fn()을 사용하지 않는다. (마찬가지로, mock/spy 등도 이번 미션에서는 사용하지 않는다.)
