--- src/Car.js
@@ -1,12 +1,0 @@
-class Car {
-    #name = "";
-    #position = 0;
-    constructor(){
-
-    }
-    move(){
-        this.position++;
-    }
-}
-
-export default Car;
