+++ src/domain/Race.js
@@ -1,0 +1,5 @@
+class Race {
+
+}
+
+export default Race;
