+++ src/domain/Winner.js
@@ -1,0 +1,5 @@
+class Winner{
+
+}
+
+export default Winner;
